Redesign itinerary implementation in modal
